Lifted state management of chargers from table component to page
